+++ shared/core/types/fsm_states.go
@@ -12,6 +12,6 @@ const (
 	StateMachineState_Consensus_SyncMode  StateMachineState = "Consensus_SyncMode"
 	StateMachineState_Consensus_Synced    StateMachineState = "Consensus_Synced"
 
-        // Used by synched validators to participate in the HotPOKT lifecycle
+	// Used by synched validators to participate in the HotPOKT lifecycle
 	StateMachineState_Consensus_Pacemaker StateMachineState = "Consensus_Pacemaker"
 )