+++ shared/node.go
@@ -146,7 +146,7 @@ func (node *Node) handleEvent(message *messaging.PocketEnvelope) error {
 	case messaging.ConsensusNewHeightEventType:
 		return node.GetBus().GetP2PModule().HandleEvent(message.Content)
 	case messaging.StateMachineTransitionEventType:
-		err_consensus := node.GetBus().GetConsensusModule().HandleMessage(message.Content)
+		err_consensus := node.GetBus().GetConsensusModule().HandleStateTransitionEvent(message.Content)
 		err_p2p := node.GetBus().GetP2PModule().HandleEvent(message.Content)
 		return multierr.Combine(err_consensus, err_p2p)
 	default: